Remove PSDEF in favor of PSL (#150)
The docs building still fails, because unfortunately the documentation
still has a cyclic dependency to QEDcore and QEDprocesses. I think the
best way forward is to remove the dependency from the literate.jl
tutorials by using self-defined custom types and removing or changing
the jldoctests that rely on QEDcore or processes. Those should probably
be moved to their respective packages.

---------

Co-authored-by: Uwe Hernandez Acosta <szabo137@users.noreply.github.com>